too fast

+++ include/philo.h
@@ -6,7 +6,7 @@
 /*   By: dzhakhan <dzhakhan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/24 12:43:51 by dzhakhan          #+#    #+#             */
-/*   Updated: 2024/09/26 16:50:17 by dzhakhan         ###   ########.fr       */
+/*   Updated: 2024/09/26 17:43:06 by dzhakhan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,10 +26,10 @@ typedef struct s_philo
 	int				r_f;
 	int				meals_eaten;
 	int				num_to_eat;
-	long			time_to_eat;
-	long			time_to_sleep;
-	long			time_to_die;
-	long			last_meal;
+	long long		time_to_eat;
+	long long		time_to_sleep;
+	long long		time_to_die;
+	long long		last_meal;
 	struct s_info	*info;
 }				t_philo;
 
@@ -38,7 +38,7 @@ typedef struct s_info
 	int				one_dead;
 	int				all_full;
 	int				philo_num;
-	long			start;
+	long long		start;
 	pthread_mutex_t	is_eating;
 	pthread_mutex_t	is_writing;
 	pthread_mutex_t	forks[200];
@@ -50,8 +50,8 @@ int		check_args(char **args);
 char	*ft_itoa(int n);
 int		ft_atoi(char const *str);
 int		ft_strcmp(char const *s1, char const *s2);
-long	ft_usleep(int ms);
-long	get_ctime(void);
+long long	ft_usleep(long long ms);
+long long	get_ctime(void);
 void	overlord(t_info *info);
 void	*routine(void *data);
 void	is_doing(char *msg, t_philo *philo);
